clean up applications-list constants
