+++ src/app/pages/applications-list-detail/applications-list-close/applications-list-close.component.ts
@@ -4,9 +4,9 @@ import { Component, OnInit, PLATFORM_ID, inject } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 
 import { AlertComponent } from '@components/alert/alert.component';
-import { APPLICATIONS_LIST_COLUMNS } from '@components/applications-list/util/applications-list.constants';
 import { ReviewConfirmComponent } from '@components/review-confirm/review-confirm.component';
 import { TableComponent } from '@components/table/table.component';
+import { APPLICATIONS_LIST_COLUMNS } from '@constants/applications-list/applications-list.constants';
 import { IF_MATCH } from '@context/concurrency-context';
 import { DateTimePipe } from '@core/pipes/dateTime.pipe';
 import { ApplicationListsApi } from '@openapi';
