+++ src/app/shared/constants/applications-list/applications-list.constants.ts
