+++ src/app/shared/constants/applications/applications.constants.ts
@@ -1,0 +1,10 @@
+export const ApplicationsColumns = [
+  { header: 'Date', field: 'date', wrap: false },
+  { header: 'Applicant', field: 'applicant' },
+  { header: 'Respondent', field: 'respondent' },
+  { header: 'Application title', field: 'title' },
+  { header: 'Fee', field: 'fee' },
+  { header: 'Resulted', field: 'resulted' },
+  { header: 'Status', field: 'status' },
+  { header: 'Actions', field: 'actions', sortable: false },
+];
