+++ src/app/shared/constants/reports/report-err.ts
@@ -1,4 +1,4 @@
-import { APPLICATIONS_LIST_CREATE_FORM_ERROR_MESSAGES } from '@components/applications-list/util/applications-list.constants';
+import { APPLICATIONS_LIST_CREATE_FORM_ERROR_MESSAGES } from '@constants/applications-list/applications-list.constants';
 import { ReportId } from '@shared-types/reports/report.types';
 
 export const REPORTS_FORM_ERROR_MESSAGES = {
